Move axTLS code to a separate namespace
In preparation for moving BearSSL into its own namespace, move axTLS to
a new namespace.  Add default "using" clause to the header for now.
